+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/config/InvertedIndexBackend.java
@@ -1,5 +1,9 @@
 package org.rcsb.strucmotif.config;
 
+import org.rcsb.strucmotif.io.codec.BucketCodec;
+import org.rcsb.strucmotif.io.codec.ColferCodec;
+import org.rcsb.strucmotif.io.codec.MessagePackCodec;
+
 /**
  * How to persist inverted index data?
  */
@@ -7,9 +11,33 @@ public enum InvertedIndexBackend {
     /**
      * Smaller files that are faster to read/write in a less standard format.
      */
-    COLFER,
+    COLFER(new ColferCodec(), ".colf"),
     /**
      * Less compact and slower MessagePack in a common serialization format.
      */
-    MESSAGE_PACK
+    MESSAGE_PACK(new MessagePackCodec(), ".msg");
+
+    private final BucketCodec bucketCodec;
+    private final String extension;
+
+    InvertedIndexBackend(BucketCodec bucketCodec, String extension) {
+        this.bucketCodec = bucketCodec;
+        this.extension = extension;
+    }
+
+    /**
+     * The implementation used by this backend.
+     * @return a {@link BucketCodec} instance
+     */
+    public BucketCodec getBucketCodec() {
+        return bucketCodec;
+    }
+
+    /**
+     * Reports the extension the index files have.
+     * @return the extension associated with this backend
+     */
+    public String getExtension() {
+        return extension;
+    }
 }